Changes to how zones are created.
Previously ComputeStacks expected to be the exclusive manager of zones
on the associated DNS Server. As a result, if you shared your DNS
infrastructure with ComputeStacks and another user created a zone, then
a separate user within ComputeStacks could create the same zone and gain
access to that. With this commit we will now enforce uniqueness at the
dns server level before creating that zone.

Admins can override this by creating the zone and unchecking the create
zone box. This is a way to import zones as an administrator.